+++ NEWS.md
@@ -12,6 +12,10 @@
 
 - Theorem environments are not defined for LaTeX output unless a `theorem` block is present (thanks, @JeffreyRacine, #291).
 
+## MINOR CHANGES
+
+- The `daemon` argument was removed from `serve_book()`, but you can still pass it to `servr::httw()` via the `...` argument.
+
 # CHANGES IN bookdown VERSION 0.3
 
 ## NEW FEATURES